Build 10 bug fix

+++ Server/db/db_object_tbl.h
@@ -27,7 +27,7 @@
         HARVESTABLE         INT,  \
         EDIBLE              INT,  \
         HARVEST_INTERVAL    INT,  \
-        EMU                 REAL, \
+        EMU                 INT, \
         EQUIPABLE_ITEM_TYPE INT,  \
         EQUIPABLE_ITEM_ID   INT)"
